Mise en place des thèmes sombre, clair et système

+++ Scripts/applications/Calculator_app.py
@@ -1,5 +1,5 @@
 #applications calculatrice
-def calul (ch1:int, ch2: int, opération :chr ): 
+def calul (ch1:float, ch2: float, opération :chr ): 
     if opération == "-":
         résultat =  ch1-ch2
     elif opération == "+":
